[development/paths] moving file and file system into win directories as they have win specific code
